requirements.txt
tinkerforge
pillow